Refactor backend into service modules
